3lab/requirements.txt
fastapi
uvicorn
celery
redis
pydantic
python-jose[cryptography]
python-multipart
websockets
aiofiles
requests
